+++ quickannotator/api/v1/tile/__init__.py
@@ -125,7 +125,7 @@ class TilePredict(MethodView):
         """
         upsert_tile(args['annotation_class_id'], args['image_id'], args['tile_id'], seen=1)
 
-        object_ref = compute_on_tile(db=db, tile_id=args['tile_id'], sleep_time=5)
+        object_ref = compute_on_tile(args['annotation_class_id'], args['image_id'], tile_id=args['tile_id'], sleep_time=5)
 
         return {'object_ref': object_ref}, 201
